feat: 修改了各个方法中参数的类型，将int类型修改为long类型
	modified:   common/src/main/java/com/scut/common/dto/response/BoardCommentDto.java
    modified:   common/src/main/java/com/scut/common/dto/response/GameDetailsDto.java
	modified:   common/src/main/java/com/scut/common/dto/response/GameForBoardDto.java
	modified:   common/src/main/java/com/scut/common/dto/response/SearchGameListDto.java

+++ common/src/main/java/com/scut/common/dto/response/GameForBoardDto.java
@@ -10,10 +10,10 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @ApiModel(description = "排行榜信息")
 public class GameForBoardDto {
-    int id; //游戏id
+    long id; //游戏id
     String name; //游戏名称
     String cover; //封面地址
     String logo; //logo地址
     String classify; //游戏分类
-    float score; //评分
+    long score; //评分
 }
